refactor(web): extract use-comment-state.ts from SpatialEditor
The comment-conversation state (commentCompose, openCommentId,
commentDrag + its settle effect, pressedCommentRef) and the geometry
helpers built on it (commentPlacementObstacles, hitTestComment,
commentById, toggleCommentCard, openCommentEditor) move verbatim into
their own hook, called right after use-scene-projection —
commentChromeBoxes is its whole hit-testing input, so no earlier
cluster needs to change.

draggedCommentId and surfaceKeyed stay in SpatialEditor.tsx as before
(unaffected bridge computation reading commentDrag from this hook's
return).

editor-state-surface.test.ts's sources glob gains
use-comment-state.ts.
