+++ apps/web/src/components/spatial-editor/SpatialEditor.tsx
@@ -68,7 +68,6 @@ import {
   BODY_LINE_HEIGHT_PX,
   COMMENT_BUBBLE_PADDING_PX,
   COMMENT_BUBBLE_RADIUS_PX,
-  commentAnchor,
   edgeLabelAnchor,
   outlineContentBox,
   placeCommentBubble,
@@ -78,29 +77,16 @@ import {
   SPATIAL_THEME_GEOMETRY,
 } from '@kamiazya/whiteboard-canvas-render'
 import { createBrowserMeasureText } from '@kamiazya/whiteboard-canvas-viewer'
-import type {
-  CanvasComment,
-  CommentThread,
-  SpatialCanvas,
-  SpatialNode,
-} from '@kamiazya/whiteboard-model'
+import type { CommentThread, SpatialCanvas, SpatialNode } from '@kamiazya/whiteboard-model'
 import { bundledFacetRegistry } from '@kamiazya/whiteboard-plugin-visual'
-import {
-  forwardRef,
-  type ReactNode,
-  useEffect,
-  useImperativeHandle,
-  useMemo,
-  useRef,
-  useState,
-} from 'react'
+import { forwardRef, type ReactNode, useImperativeHandle, useMemo, useRef, useState } from 'react'
 import { writeLastTool } from '@/lib/initial-tool'
 import type { ResolvedTheme } from '../../hooks/useThemeMode.js'
 import { parseClipboardText } from '../../lib/clipboard-fragment.js'
 import { hapticTick } from '../../lib/haptics.js'
 import { hasCoarsePointer } from '../../lib/platform.js'
 import type { BoxMove } from './align.js'
-import { CanvasContextMenu, type CommentComposeState } from './CanvasContextMenu.js'
+import { CanvasContextMenu } from './CanvasContextMenu.js'
 import { CommentDragLayer } from './CommentDragLayer.js'
 import { CommentThreadCard } from './CommentThreadCard.js'
 import { ConnectOverlay } from './ConnectOverlay.js'
@@ -163,6 +149,7 @@ import {
 } from './ToolPalette.js'
 import { useCanvasReplacement } from './use-canvas-replacement.js'
 import { useClipboardActions } from './use-clipboard-actions.js'
+import { useCommentState } from './use-comment-state.js'
 import { useDragLayers } from './use-drag-layers.js'
 import { useEditSessionState } from './use-edit-session-state.js'
 import { useEditorKeyboard } from './use-editor-keyboard.js'
@@ -714,107 +701,20 @@ export const SpatialEditor = forwardRef<SpatialEditorHandle, SpatialEditorProps>
     )
     const { keyed, edgePaths, commentChromeBoxes, selectionMembers, selectionBox, minimapNodes } =
       useSceneProjection({ scene, bounds, boxes, canvas, theme, selectedId, extraIds })
-    /**
-     * What a comment's bubble is placed around — the same obstacle set
-     * canvas-render's placer sees for it: every node that is not a group
-     * frame, plus the bubbles of the comments BEFORE it in document order
-     * (all of them for a comment about to be created, which goes last). The
-     * editor's draft and its drag preview both place through this, so a
-     * bubble opens, drags and settles in one spot.
-     */
-    const commentPlacementObstacles = (beforeCommentId?: string): BoundingBox[] => {
-      const out: BoundingBox[] = canvasRef.current.nodes
-        .filter((node) => node.type !== 'group')
-        .map((node) => ({ x: node.x, y: node.y, w: node.width, h: node.height }))
-      for (const entry of commentChromeBoxes) {
-        if (entry.part !== 'bubble') continue
-        if (entry.commentId === beforeCommentId) break
-        out.push(entry.bbox)
-      }
-      return out
-    }
-    const hitTestComment = (point: Point): string | undefined => {
-      for (let i = commentChromeBoxes.length - 1; i >= 0; i -= 1) {
-        const entry = commentChromeBoxes[i]
-        if (entry === undefined) continue
-        const { bbox } = entry
-        if (
-          point.x >= bbox.x &&
-          point.x <= bbox.x + bbox.w &&
-          point.y >= bbox.y &&
-          point.y <= bbox.y + bbox.h
-        ) {
-          return entry.commentId
-        }
-      }
-      return undefined
-    }
-    const commentById = (id: string): CanvasComment | undefined =>
-      canvasRef.current['x-whiteboard']?.comments?.find((entry) => entry.id === id)
-    /**
-     * Opens a conversation in place, or shuts the one already open. Pressing
-     * the comment whose card is up is how it closes without hunting for the
-     * ×, which is the gesture people try first.
-     */
-    const toggleCommentCard = (commentId: string): void => {
-      setOpenCommentId((current) => (current === commentId ? null : commentId))
-    }
-    /** Opens the compose bubble over an existing comment, pre-filled, to rewrite its text. */
-    const openCommentEditor = (comment: CanvasComment) => {
-      setCommentCompose({
-        point: commentAnchor(comment, canvasRef.current),
-        editing: { id: comment.id, initialText: comment.text },
-      })
-    }
-    const [commentCompose, setCommentCompose] = useState<CommentComposeState | null>(null)
-    /**
-     * The conversation opened in place on the canvas — at most one, because
-     * a card is a place to read and answer ONE thread, and a canvas of
-     * simultaneously open cards is the comments rail with worse layout.
-     */
-    const [openCommentId, setOpenCommentId] = useState<string | null>(null)
-    /**
-     * The comment a press landed on, read back at the release. A press that
-     * never travelled opens the card; one that travelled was a pin drag and
-     * the drag branch owns it.
-     */
-    const pressedCommentRef = useRef<string | null>(null)
-    /**
-     * A point-anchored comment's pin being dragged: the comment as it was
-     * at the press (its stored anchor), the press point, and the live
-     * pointer. Kept beside the gesture machine rather than in it — see
-     * CommentDragLayer for why — so the node machinery (snapping, carried
-     * sets, live edges) never has to learn what a comment is.
-     */
-    const [commentDrag, setCommentDrag] = useState<{
-      readonly comment: CanvasComment
-      readonly startPoint: Point
-      readonly live: Point | null
-      /** The obstacle set at the press, so the preview places like the committed chrome. */
-      readonly obstacles: readonly BoundingBox[]
-      /**
-       * Set on release: the anchor the move was committed at. The drag then
-       * SETTLES rather than ending — the preview stays up, and the committed
-       * copy stays out of the surface, until the committed scene carries the
-       * comment at this anchor (a worker round trip later on a large canvas).
-       * Ending at the release instead showed the old copy for a frame and
-       * animated it to the new anchor.
-       */
-      readonly dropped: Point | null
-    } | null>(null)
-    useEffect(() => {
-      if (commentDrag?.dropped == null) return
-      const { dropped } = commentDrag
-      const pin = commentChromeBoxes.find(
-        (entry) => entry.commentId === commentDrag.comment.id && entry.part === 'pin',
-      )
-      const arrived =
-        pin !== undefined &&
-        pin.bbox.x + pin.bbox.w / 2 === dropped.x &&
-        pin.bbox.y + pin.bbox.h / 2 === dropped.y
-      // Gone (removed underneath the drag) settles too: nothing to wait for.
-      if (arrived || commentById(commentDrag.comment.id) === undefined) setCommentDrag(null)
-    }, [commentDrag, commentChromeBoxes])
+    const {
+      commentPlacementObstacles,
+      hitTestComment,
+      commentById,
+      toggleCommentCard,
+      openCommentEditor,
+      commentCompose,
+      setCommentCompose,
+      openCommentId,
+      setOpenCommentId,
+      pressedCommentRef,
+      commentDrag,
+      setCommentDrag,
+    } = useCommentState({ canvasRef, commentChromeBoxes })
     // The committed surface without the comment in flight (see
     // keyedWithoutPrefix for why it leaves rather than hides).
     const draggedCommentId = commentDrag?.comment.id
