+++ apps/web/src/components/spatial-editor/editor-state-surface.test.ts
@@ -38,6 +38,7 @@ import { assertScannedLedger } from '../../test-utils/coverage-ledger.js'
 const sources = import.meta.glob(
   [
     './SpatialEditor.tsx',
+    './use-comment-state.ts',
     './use-edit-session-state.ts',
     './use-editor-measurements.ts',
     './use-file-seam-scene.ts',
